requirements.txt
asgiref==3.3.1
astroid==2.4.2
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.26
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.0
pylint==2.6.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.3
wrapt==1.12.1
